+++ packages/examples/basic/components/user/Container.js
@@ -4,7 +4,13 @@ import { Paper, FormControl, FormLabel } from '@material-ui/core';
 import ColorPicker from 'material-ui-color-picker';
 import React from 'react';
 
-export const Container = ({ background, padding, children, outline }) => {
+export const Container = ({
+  background,
+  padding,
+  children,
+  outline,
+  ...props
+}) => {
   const {
     connectors: { connect, drag },
     id: containerNodeId,
@@ -38,6 +44,7 @@ export const Container = ({ background, padding, children, outline }) => {
 
   return (
     <Paper
+      {...props}
       ref={(ref) => connect(drag(ref))}
       style={{
         margin: '5px 0',
@@ -66,6 +73,7 @@ export const ContainerSettings = () => {
       <FormControl fullWidth={true} margin="normal" component="fieldset">
         <FormLabel component="legend">Background</FormLabel>
         <ColorPicker
+          name="background-color"
           value={background}
           onChange={(color) => {
             setProp((props) => (props.background = color), 500);