+++ packages/examples/basic/pages/index.js
@@ -39,12 +39,28 @@ export default function App() {
         <Grid container spacing={5} style={{ paddingTop: '10px' }}>
           <Grid item xs>
             <Frame>
-              <Element canvas is={Container} padding={5} background="#eeeeee">
-                <Card />
-                <Button text="Click me" size="small" />
-                <Text fontSize={20} text="Hi world!" />
-                <Element canvas is={Container} padding={6} background="#999999">
-                  <Text size="small" text="It's me again!" />
+              <Element
+                canvas
+                is={Container}
+                padding={5}
+                background="#eeeeee"
+                data-cy="root-container"
+              >
+                <Card data-cy="frame-card" />
+                <Button text="Click me" size="small" data-cy="frame-button" />
+                <Text fontSize={20} text="Hi world!" data-cy="frame-text" />
+                <Element
+                  canvas
+                  is={Container}
+                  padding={6}
+                  background="#999999"
+                  data-cy="frame-container"
+                >
+                  <Text
+                    size="small"
+                    text="It's me again!"
+                    data-cy="frame-container-text"
+                  />
                 </Element>
               </Element>
             </Frame>