+++ packages/slate/src/extend/editable/useCraftStateSync.tsx
@@ -2,7 +2,7 @@
 import { useEditor } from '@craftjs/core';
 import isEqual from 'lodash/isEqual';
 import { useEffect, useRef } from 'react';
-import { Editor } from 'slate';
+import { Editor, Node } from 'slate';
 import { useSlate } from 'slate-react';
 
 import { SlateElement } from '../../interfaces';
@@ -30,7 +30,7 @@ export const useCraftStateSync = () => {
   const slateSelectionRef = useRef(null);
   const isCraftOverriding = useRef(false);
 
-  const { actions, craftSlateState } = useEditor((state, query) => ({
+  const { query, actions, craftSlateState } = useEditor((state, query) => ({
     craftSlateState: getSlateStateFromCraft(id, query),
   }));
 
@@ -98,21 +98,24 @@ export const useCraftStateSync = () => {
   }, [slateSelection]);
 
   useEffect(() => {
-    const { apply, onChange } = slateEditor;
+    const { apply } = slateEditor;
 
     slateEditor.onChange = () => {
-      onChange();
-
-      // Don't update Craft state if only selection operations had been performed
-      if (slateEditor.operations.every((op) => op.type === 'set_selection')) {
-        return;
-      }
-
       let actionCreator = actions.history.throttle(500);
 
       if (isCraftOverriding.current) {
         actionCreator = actions.history.merge();
       }
+
+      const slateNodeIds = Array.from(Node.descendants(slateEditor)).map(
+        ([nodeId]) => nodeId.id
+      );
+      const currentSlateNodeIds = query.node(id).descendants(true);
+
+      // Get deleted slate node ids
+      const deletedNodeIds = currentSlateNodeIds.filter(
+        (id) => slateNodeIds.includes(id) === false
+      );
 
       actionCreator.setState((state) => {
         const slateState = slateEditor.children;
@@ -136,19 +139,9 @@ export const useCraftStateSync = () => {
         state.nodes[id].data.custom.selection = slateSelection;
 
         // Remove any deleted Slate nodes from the Craft state
-        slateEditor.operations
-          .filter((op) => op.type === 'remove_node')
-          .forEach((op) => {
-            if (!op.node) {
-              return;
-            }
-
-            if (!state.nodes[op.node.id]) {
-              return;
-            }
-
-            delete state.nodes[op.node.id];
-          });
+        deletedNodeIds.forEach(
+          (deletedNodeId) => delete state.nodes[deletedNodeId]
+        );
       });
     };