+++ packages/utils/CHANGELOG.md
@@ -3,6 +3,88 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 0.2.0-alpha.23 (2021-05-14)
+
+
+### Features
+
+* **slate:** improve speed of slate/craft integration ([#241](https://github.com/prevwong/craft.js/issues/241)) ([763274c](https://github.com/prevwong/craft.js/commit/763274c4787130675127703d333ea63ee8482fb1))
+
+
+
+
+
+# 0.2.0-alpha.22 (2021-04-26)
+
+
+### Bug Fixes
+
+* orphaned linked nodes ([#236](https://github.com/prevwong/craft.js/issues/236)) ([65450ea](https://github.com/prevwong/craft.js/commit/65450ea97a0bbac4efc51d0306025bd07a5165bb))
+
+
+
+
+
+# 0.2.0-alpha.21 (2021-03-22)
+
+**Note:** Version bump only for package @craftjs/utils
+
+
+
+
+
+# 0.2.0-alpha.20 (2021-03-22)
+
+
+### Bug Fixes
+
+* remove slate node ([#217](https://github.com/prevwong/craft.js/issues/217)) ([9f7ec97](https://github.com/prevwong/craft.js/commit/9f7ec97e88825460054f8880bb4b5900c2776721))
+
+
+
+
+
+# 0.2.0-alpha.19 (2021-03-16)
+
+
+### Features
+
+* additional indicator options ([#212](https://github.com/prevwong/craft.js/issues/212)) ([0d9c6e5](https://github.com/prevwong/craft.js/commit/0d9c6e59a8725ba600d177ad5a9beec0434e640c))
+
+
+
+
+
+# 0.2.0-alpha.18 (2021-03-11)
+
+**Note:** Version bump only for package @craftjs/utils
+
+
+
+
+
+# 0.2.0-alpha.17 (2021-03-11)
+
+
+### Bug Fixes
+
+* lint ([3750b11](https://github.com/prevwong/craft.js/commit/3750b11dbe27b477820dac8c940bcb46981e6956))
+
+
+
+
+
+# 0.2.0-alpha.16 (2021-02-18)
+
+
+### Bug Fixes
+
+* handlers not being disabled ([#189](https://github.com/prevwong/craft.js/issues/189)) ([f9c2274](https://github.com/prevwong/craft.js/commit/f9c227415d3bc4ef0490437b16a862234c15d7f7))
+
+
+
+
+
 # 0.2.0-alpha.15 (2021-02-11)
 
 **Note:** Version bump only for package @craftjs/utils